Added the Editor
Editor is still in progress. The file contains 2 apps. First is the
showdown side with its files. Inside the src of that there is a file
called "latest" where the markitup editor is. We want to markitup
editor's toolbar to be integrated into the showdown app to easily be
able to add the basic commands of bold italics etc.